Update node e2e (#226)
This refactors Node end-to-end tests so that they are aligned with the setup we use in other repositories. Namely, environments are used to drive a test matrix, and handle environmnent variables cleanly.

Note that in the case of some environments, the current library implementation does not work. As this is known, the associated tests are skipped until the issue is resolved.