SDK-3368: Feature flag problem details tests
The end-to-end tests need to run against older supported versions of
ESS, not all of which support problem details.

By default, only the response status and title are checked, which are
provided by default against any HTTP response. Problem details and
instance are only checked if the environment variable E2E_TEST_FEATURE_PROBLEM_DETAILS
is set to true.